re_intuit_start(): change definition of ml_anch
The ml_anch variable is supposed to indicate that a multi-line match is
possible, i.e. that the regex is anchored to a \n.

Currently we just do

    ml_anch = (prog->intflags & PREGf_ANCH_MBOL);

However, MBOL is also set on /.*..../; the two cases are distinguished by
adding the PREGf_IMPLICIT flag too.

So at the moment we have lots of tests along the lines of

    if (ml_anch && !(prog->intflags & PREGf_IMPLICIT))

Simplify this by adding the IMPLICIT condition when initially calculating
ml_anch, so there's no need to keep testing for it later. This also means
that ml_anch actually means what it says now.