re_intuit_start(): don't unset MBOL on uselessness
When BmUSEFUL() for a pattern goes negative, we unset ~RXf_USE_INTUIT.

A bug fix in 2010 (c941595168) made this part of the code also remove
the RXf_ANCH_MBOL at the same time, if PREGf_IMPLICIT was set.

However, this was really just working round a bug in regexec_flags().
Once intuit was disabled, regexec_flags() would then start taking the
buggy code path.

This buggy code path was separately fixed in 2012 by 21eede782, so there's
no longer any need to remove this flag. So don't.