Remove MM_TEST_ROOT feature from ExtUtils::Install tests.
This feature depended on test directories created on the fly being
in a predictable location, but as of 80af860fb39a3c6 they no longer
are.  VMS systems released in the last 15 years or so don't even
need the feature, so the best path forward is to get rid of it.

A corresponding change was made in ExtUtils::MakeMaker 6.79_01 but
ExtUtils::Install maintains its own fork of the same test libraries.