[perl #121484] /m causing false negative
My recent commit d0d4464849e2b30aee8 in re_intuit_start() reduced the
scope of a 'skip if multiline' check, so that certain optimisations
weren't being unnecessarily skipped. Unfortunately it didn't reduce the
scope enough, so a vital slen-- was being skipped in the
SvTAIL-but-don't-fail case.

This commit just moves the !multiline test further down, and updates the
commentary and condition formatting a bit.