Add perlbug -E option to specify report encoding.
But default to UTF-8 when no encoding is specified. We actually use
the internal :utf8 layer so we can survive without the Encode module
if necessary, such as under miniperl.

The encoding applies to report bodies created ahead of time and
supplied with -f and also those for which we provide a template and
the user fills out the template in an editor.  It is assumed the
user's editor can handle the encoding supplied and will leave it
unchanged.

The body is converted to UTF-8 on output for mail transport or when
saving the entire report to disk.  If there are any attachments, we
switch the output layer to :raw for those as that's how we read them
in and we make no assertion about what encoding(s) they are in.

Since we are now treating the body as text rather than binary, we
don't need to add the :crlf layer explicitly for Windows.

In order to switch output layers between the report body and any
attachments, some refactoring of the functions that assemble the
different components of the mail message was necessary.  Also, we now
store mail headers in an array rather than a hash to avoid a hash
order bug that was probably always present but rarely if ever
triggered until we added more headers to handle attachments.