in the share_hek logic test flags before testing strings
Normally we test the flags after as the flag check rarely fires. However
in PL_strtab certain rules are a little different, and we have a higher chance
of collissions, so we check the flag first.

Also commented out the various places we do this so people dont wonder
about the discrepancy in the future.