Coverity 45351 - Refactor to skip a loop iteration
The loop in this case will remove an element from a list,
so it keeps a pointer to the previous element.  Before this
change, the pointer to the previous element was only initialized
at the end of the first pass, so Coverity flagged the code
as potentially using a NULL pointer.  It's a false positive
because the first element in the list is always a pushmark, so
the branch that actually used that pointer would never be hit
while it's still NULL, but changing the code to explicitly set
lasto before entering the loop (and asserting that its a PUSHMARK)
makes the logic clearer to both humans and Coverity.